# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Alpha server built with Go and Chi",
        "title": "StayHealthy Swagger Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/waistline/{id}": {
            "get": {
                "description": "Get user waist line",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "measurements waistline"
                ],
                "summary": "Get user waist line",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "WaistLine",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/measurement.WaistLine"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete user waist line",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "measurements waistline"
                ],
                "summary": "Delete user waist line",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "WaistLine",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/measurement.WaistLine"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update user waist line",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "measurements waistline"
                ],
                "summary": "Update user waist line",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "WaistLine",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/measurement.WaistLine"
                        }
                    }
                }
            }
        },
        "/water": {
            "get": {
                "description": "Get user water intake list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "measurements water"
                ],
                "summary": "Get user water intake list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/measurement.WaterIntake"
                        }
                    }
                }
            },
            "post": {
                "description": "Insert user water intake",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "measurements water"
                ],
                "summary": "Insert user water intake",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Water",
                        "name": "w",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/measurement.WaterIntake"
                        }
                    }
                }
            }
        },
        "/water/{id}": {
            "get": {
                "description": "Get user water intake",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "measurements water"
                ],
                "summary": "Get user water intake",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Water",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/measurement.WaterIntake"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete user water intake",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "measurements water"
                ],
                "summary": "Delete user water intake",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Water",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/measurement.WaterIntake"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update user water intake",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "measurements water"
                ],
                "summary": "Update user water intake",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Water",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/measurement.WaterIntake"
                        }
                    }
                }
            }
        },
        "/weight": {
            "post": {
                "description": "Insert user weight",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "measurements weight"
                ],
                "summary": "Insert user weight",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Weight",
                        "name": "w",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/measurement.Weight"
                        }
                    }
                }
            }
        },
        "/weight/{id}": {
            "get": {
                "description": "Get user weight",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "measurements weight"
                ],
                "summary": "Get user weight",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Weight",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/measurement.Weight"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete user weight",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "measurements weight"
                ],
                "summary": "Delete user weight",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Weight ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/measurement.Weight"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update user weight",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "measurements weight"
                ],
                "summary": "Update user weight",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Weight ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/measurement.Weight"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "measurement.WaistLine": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": ""
                },
                "quantity": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "measurement.WaterIntake": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": ""
                },
                "quantity": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "measurement.Weight": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": ""
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "weight_value": {
                    "type": "number"
                }
            }
        },
...
    }
}